History eliminates passed-in state
We don't need to push tab switches into the history, since executing a "back" button-like command should theoretically take you back to the page before the TabNav rather than the previous tab you were on (maybe this is just me, but I *hate* when websites use slideshows and then hitting back after you've clicked through takes you to a previous image--that's the analogy I'm working off of). Also, pushing to history disrupts any state passed in via links (currently, the edit record page is the only thing that uses a TabNav and it relies on state to keep track of which child in which report you're actually editing).